удалил упоминание о классе AbonCard. Изменил название Client на Person
