+++ src/main/java/ru/pork/util/PersonManager.java
@@ -2,18 +2,18 @@ package ru.pork.util;
 
 import com.sun.xml.internal.ws.handler.HandlerException;
 import org.hibernate.*;
-import ru.pork.model.Client;
+import ru.pork.model.Person;
 import ru.pork.model.Contracts;
 import ru.pork.servlet.DatabaseConfigurator;
 
 import java.util.*;
 
 
-public class ClientManager {
+public class PersonManager {
     private SessionFactory factory;
 
 
-    public ClientManager() {
+    public PersonManager() {
         try {
             factory= DatabaseConfigurator.getSessionFactory();
         } catch (HandlerException hex ) {
@@ -24,16 +24,16 @@ public class ClientManager {
 
 
 
-    public boolean addClient(Client client) {
+    public boolean addClient(Person person) {
         Transaction tx;
         Session session=factory.openSession();
 
         tx=session.beginTransaction();
         try {
-            if (client.getProgram()==null) {
+            if (person.getProgram()==null) {
                 return false;
             } else {
-                session.save(client);
+                session.save(person);
                 tx.commit();
                 return true;
             }
@@ -46,15 +46,15 @@ public class ClientManager {
         }
     }
 
-    public List<Client> listClients() {
+    public List<Person> listClients() {
         Transaction tx;
         Session session=factory.openSession();
 
-        ArrayList<Client> list = new ArrayList<Client>();
+        ArrayList<Person> list = new ArrayList<Person>();
         tx=session.beginTransaction();
         try {
-            List lists = session.createQuery("FROM Client ").list();
-            for (Iterator<Client> iterator = lists.iterator(); iterator.hasNext(); ) {
+            List lists = session.createQuery("FROM Person").list();
+            for (Iterator<Person> iterator = lists.iterator(); iterator.hasNext(); ) {
                 list.add(iterator.next());
             }
             tx.commit();
@@ -73,8 +73,8 @@ public class ClientManager {
         Session session=factory.openSession();
         tx=session.beginTransaction();
         try {
-            Client client=session.get(Client.class, id);
-            session.delete(client);
+            Person person =session.get(Person.class, id);
+            session.delete(person);
             tx.commit();
             session.close();
             return true;
@@ -91,9 +91,9 @@ public class ClientManager {
         Session session = factory.openSession();
         tx = session.beginTransaction();
         try {
-            Client client = session.get(Client.class, id);
-            client.getContractses().add(contracts);
-            session.update(client);
+            Person person = session.get(Person.class, id);
+            person.getContracts().add(contracts);
+            session.update(person);
             tx.commit();
             session.close();
             return true;
@@ -105,7 +105,7 @@ public class ClientManager {
         }
     }
 
-    public Client findClient(long phon) {
+    public Person findClient(long phon) {
         Transaction tx;
         Session session=factory.openSession();
         tx=session.beginTransaction();
@@ -113,12 +113,12 @@ public class ClientManager {
             Query q=session.createQuery("from Client  C where C.phone=:phone");
             q.setParameter("phone", phon);
             List clients=q.list();
-            for (Iterator<Client> iterator=clients.iterator();iterator.hasNext();) {
-                Client client = iterator.next();
-                if (client.getPhone()==phon) {
+            for (Iterator<Person> iterator=clients.iterator();iterator.hasNext();) {
+                Person person = iterator.next();
+                if (person.getPhone()==phon) {
                     tx.commit();
                     session.close();
-                    return client;
+                    return person;
                 }
             }
             tx.commit();
@@ -131,12 +131,12 @@ public class ClientManager {
             return null;
         }
     }
-    public Client findClient(int id) {
+    public Person findClient(int id) {
         Transaction tx;
         Session session=factory.openSession();
         tx=session.beginTransaction();
         try {
-            Client cl=session.get(Client.class,id);
+            Person cl=session.get(Person.class,id);
             tx.commit();
             session.close();
             return  cl;
