--- src/main/resources/AbonCard.hbm.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!DOCTYPE hibernate-mapping PUBLIC
-        "-//Hibernate/Hibernate Mapping DTD//EN"
-        "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
-<hibernate-mapping>
-    <class name="ru.pork.model.AbonCard" table="AbonCard">
-        <meta attribute="class-description">
-            This class describe Abonement Card
-        </meta>
-        <id name="id" type="int" column="id">
-            <generator class="native"/>
-        </id>
-        <!--<property name="cardType" column="cardType" type="int"/>-->
-        <property name="description" column="description" type="string"/>
-        <!--<property name="expiration_begin" column="expiration_begin" type="date"/>-->
-        <!--<property name="expiration_end" column="expiration_end" type="date"/>-->
-        <property name="price" column="price" type="int"/>
-        <!--<property name="typeOfDiscount" column="typeOfDiscout" type="int"/>-->
-        <!--<property name="discountPercent" column="discountPercent" type="int"/>-->
-    </class>
-</hibernate-mapping>
