--- src/main/web/abonements_add.jsp
@@ -1,25 +1,0 @@
-
-<%--
-  Created by IntelliJ IDEA.
-  User: santa
-  Date: 29.02.16
-  Time: 16:59
-  To change this template use File | Settings | File Templates.
---%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-    <title>Абонементы</title>
-</head>
-<body>
-<%--<center>--%>
-  <form method="get" action="aboncard">
-    Уникальное название абонемента: <input type="text" name="description"/><br/>
-    Цена абоненмента: <input type="number" name="price"/></br>
-    <input type="submit" name="Создать"></br>
-  </form>
-<%--</center>--%>
-<a href="welcome.jsp">Назад</a>
-
-</body>
-</html>
